removed something minor and made 2 functions
the functions check if wifi and celltower with roomid and id and return
a boolean value, they return true even if there is more than one match
but they goes into a different branch. they are a little heavy so we
might have do something else if we have to use them alot.